+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
   [#3348](https://github.com/pulumi/pulumi/pull/3348)
 - Improve startup time performance by 0.5-1s by checking for a newer CLI release in parallel.
   [#3441](https://github.com/pulumi/pulumi/pull/3441)
+- Add an experimental `pulumi watch` command. [#3391](https://github.com/pulumi/pulumi/pull/3391)
 
 ## 1.4.1 (2019-11-01)